Add files via upload
BeamScheduling:
Download h3 library, terminal input: pip install h3==4.0.0b2 -i http://pypi.douban.com/simple --trusted-host pypi.douban.com
The reason for the difference in output results each time: the output of the following functions is the set() category, and the order is different. The output results of the same input of the h3.grid_ring function are in different order, but the content is the same, and the order of cell_to_children and grid_disk is also different.
You can change global parameters in 'BeamScheduling.py', in which the corresponding parameters are commented.